AKA how to be faster than ripgrep ;)

Fixes https://github.com/microsoft/vscode-js-debug/issues/1149
Fixes https://github.com/microsoft/vscode/issues/167911

This does some overall system optimization for how the breakpoint
predictor works. It achieves performance comparable to the VS Code API
ripgrep globbing, with plain JS. And by using plain JS, we can do more
optimizations in a future PR.

Previously, we had an `ISearchStrategy` type that globbed for all
sourcemap URIs in a target folder, and emitted those. Then the breakpoint
predictor had an mtime-based cache to decide whether it would parse the
sourcemap or not.

This moves the caching responsibility into `ISearchStrategy`, which then
takes a two-phase approach to processing files, allowing it to cache the
intermediate result the breakpoint predictor previously cached itself.

This means that:

- Files never need to be read if their mtime doesn't change
- When starting a new child session, we optimize further and don't stat
  compiled files that didn't previously reference the sourcefile. (We
  still stat all directories to see if there are new files to pick up)

Some performance numbers:

- vscode#167911 time on my computer going from 33s -> 16s
- typescript: time to run mocha tests 52s -> 22s (without outFiles configured)

The new search strategy can be disabled by setting
`enableTurboSourcemaps: false` in your launch.json.

![](https://media4.giphy.com/media/cRH5deQTgTMR2/giphy.gif?cid=ecf05e47r7gdpd1vzjtyo72z0ymmob3ltfjz7ts4cs64n18i&rid=giphy.gif&ct=g)